app/services/storage.py
"""
File storage service for audio files using Docker volumes

TODO Phase 2:
- Implement secure file upload
- Add file streaming with range support
- Implement file deletion
- Add file integrity verification
- Implement virus scanning integration
"""

import os
import hashlib
from pathlib import Path
from typing import BinaryIO, AsyncGenerator, Optional
import aiofiles
import magic

from app.core.config import settings
from app.core.logging import get_logger
from app.core.exceptions import FileUploadError

logger = get_logger(__name__)


class StorageService:
    """
    Service for managing audio file storage

    TODO Phase 2:
    - Implement file encryption at rest
    - Add automatic file cleanup
    - Implement storage quotas
    - Add file deduplication
    - Implement backup/restore
    """

    def __init__(self):
        self.base_path = Path(settings.AUDIO_STORAGE_PATH)
        self.max_size = settings.MAX_UPLOAD_SIZE
        self.allowed_formats = settings.ALLOWED_AUDIO_FORMATS
        self.chunk_size = settings.CHUNK_SIZE

    async def save_audio_file(
        self,
        file: BinaryIO,
        filename: str,
        user_id: int,
    ) -> dict:
        """
        Save uploaded audio file to Docker volume

        Args:
            file: File object to save
            filename: Original filename
            user_id: ID of user uploading file

        Returns:
            Dictionary with file metadata:
            - file_path: Relative path to file
            - file_size: Size in bytes
            - mime_type: Detected MIME type
            - file_hash: SHA-256 hash
            - duration: Audio duration in seconds (if detectable)

        TODO Phase 2:
        - Validate file format
        - Generate secure filename
        - Create user directory if not exists
        - Stream file to disk
        - Calculate file hash
        - Extract audio metadata
        - Verify file integrity
        - Implement virus scanning
        - Log file upload

        Raises:
            FileUploadError: If upload fails
        """
        pass

    async def stream_audio_file(
        self,
        file_path: str,
        start_byte: Optional[int] = None,
        end_byte: Optional[int] = None,
    ) -> AsyncGenerator[bytes, None]:
        """
        Stream audio file with optional range support

        Args:
            file_path: Path to file relative to storage base
            start_byte: Optional start byte for range request
            end_byte: Optional end byte for range request

        Yields:
            Chunks of file data

        TODO Phase 2:
        - Validate file exists
        - Validate range parameters
        - Stream file in chunks
        - Support HTTP 206 Partial Content
        - Log file access
        - Implement bandwidth throttling

        Raises:
            FileNotFoundError: If file doesn't exist
            ValueError: If invalid range
        """
        pass

    async def delete_audio_file(self, file_path: str) -> None:
        """
        Delete audio file from storage

        Args:
            file_path: Path to file relative to storage base

        TODO Phase 2:
        - Validate file exists
        - Delete file securely (overwrite before delete)
        - Log file deletion
        - Handle errors gracefully

        Raises:
            FileNotFoundError: If file doesn't exist
        """
        pass

    async def get_file_info(self, file_path: str) -> dict:
        """
        Get metadata about stored file

        Args:
            file_path: Path to file relative to storage base

        Returns:
            Dictionary with file info:
            - size: File size in bytes
            - mime_type: MIME type
            - last_modified: Last modification timestamp

        TODO Phase 2:
        - Get file stats
        - Detect MIME type
        - Calculate hash if needed
        - Extract audio metadata
        """
        pass

    def _generate_secure_filename(self, original_filename: str, user_id: int) -> str:
        """
        Generate secure, unique filename

        Args:
            original_filename: Original uploaded filename
            user_id: ID of user

        Returns:
            Secure filename

        TODO Phase 2:
        - Extract file extension
        - Validate extension
        - Generate UUID for uniqueness
        - Include timestamp
        - Prevent path traversal
        """
        pass

    def _validate_file_format(self, file: BinaryIO) -> str:
        """
        Validate file format using magic bytes

        Args:
            file: File object to validate

        Returns:
            Detected MIME type

        TODO Phase 2:
        - Read magic bytes
        - Detect MIME type
        - Validate against allowed formats
        - Don't trust file extension

        Raises:
            FileUploadError: If format not allowed
        """
        pass

    async def _calculate_file_hash(self, file_path: Path) -> str:
        """
        Calculate SHA-256 hash of file

        Args:
            file_path: Path to file

        Returns:
            Hex-encoded SHA-256 hash

        TODO Phase 2:
        - Read file in chunks
        - Calculate hash
        - Use for deduplication and integrity
        """
        pass

    async def _extract_audio_metadata(self, file_path: Path) -> dict:
        """
        Extract metadata from audio file

        Args:
            file_path: Path to audio file

        Returns:
            Dictionary with metadata:
            - duration: Duration in seconds
            - bitrate: Bitrate in kbps
            - channels: Number of channels
            - sample_rate: Sample rate in Hz

        TODO Phase 2:
        - Use mutagen or similar library
        - Extract duration, bitrate, etc.
        - Handle different audio formats
        - Handle corrupt files gracefully
        """
        pass


# Singleton storage service instance
storage_service = StorageService()
