pyproject.toml
[project]
name = "dictat"
version = "0.1.0"
description = "Medical dictation service with transcription workflow management"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0" }
authors = [
    { name = "Dictat Team" }
]
keywords = ["medical", "dictation", "transcription", "fastapi", "gdpr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "argon2-cffi>=23.1.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Redis & Caching
    "redis>=5.0.1",
    "hiredis>=2.3.2",

    # Background Tasks
    "celery>=5.3.4",
    "arq>=0.25.0",

    # WebSocket
    "websockets>=12.0",

    # OPA Integration
    "httpx>=0.26.0",
    "aiohttp>=3.9.1",

    # File Handling
    "python-magic>=0.4.27",
    "aiofiles>=23.2.1",

    # Email
    "fastapi-mail>=1.4.1",

    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "python-json-logger>=2.0.7",

    # Data Validation
    "email-validator>=2.1.0",
    "markdown>=3.5.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "faker>=22.0.0",
    "factory-boy>=3.3.0",

    # Code Quality
    "ruff>=0.1.14",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Security
    "bandit>=1.7.6",
    "safety>=3.0.0",

    # Performance Testing
    "locust>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing --cov-report=html --cov-report=xml"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "security: marks tests as security tests",
    "gdpr: marks tests as GDPR compliance tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 80

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "fastapi_mail.*",
    "celery.*",
    "arq.*",
    "magic.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]
